+++ Src/Web/Web.Net45/Implementation/HttpRequestExtensions.cs
@@ -3,6 +3,7 @@
     using System;
     using System.Collections.Specialized;
     using System.Web;
+    using Microsoft.ApplicationInsights.Common;
 
     /// <summary>
     /// HttpRequest Extensions.
@@ -16,7 +17,8 @@
 
         public static string UnvalidatedGetHeader(this HttpRequest httpRequest, string headerName)
         {
-            return httpRequest.Unvalidated.Headers[headerName];
+            string value = httpRequest.Unvalidated.Headers[headerName];
+            return StringUtilities.EnforceMaxLength(value, InjectionGuardConstants.RequestHeaderMaxLength);
         }
 
         public static Uri UnvalidatedGetUrl(this HttpRequest httpRequest)